requirements.txt
langchain>=0.0.150 
langchain-openai>=0.0.150 
pydantic>=1.10.0 
requests>=2.25.0 
numpy>=1.21.0
pandas>=1.3.0 
scikit-learn>=0.24.0
tqdm>=4.62.0 
python-dotenv>=0.19.0 
langchain-core>=0.0.150 
openai>=0.27.0  
langchain-google-generative-ai>=0.0.150 
google-generative-ai>=0.3.0 
langchain-huggingface>=0.0.150 
transformers>=4.21.0 
huggingface-hub>=0.8.0
torch>=1.10.0
accelerate>=0.12.0

